+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * disregards the open/closedness of the boundary
 * fix `openclosed` and `closedopen`
 * rename `imapS -> imapLev` and `itraverseS -> itraverseLev` following `Suspension -> Levitated`
+* `(...)` as a value-level analogue to `interval`
 
 ## 0.2.0.1 - 2022-05-30