+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * rename `imapS -> imapLev` and `itraverseS -> itraverseLev` following `Suspension -> Levitated`
 * `(...)` as a value-level analogue to `interval`
 * `Data.Interval.isSubsetOf`
+* `Data.Interval.Borel.isSubsetOf`
 * `Data.Interval.Layers.difference`
 
 ## 0.2.0.1 - 2022-05-30